Don't allow editing MonitoredVariable from table
This was causing it to load to slow as it was creating a dropdown for PVName for every row (there are 100 rows per page) that had ~23000 items in it. Django did this really inefficiently.

You can still edit the entry by first selecting the ID. I also removed the custom form for adding and editing, this allows the dropdown to be created there but since it is just for a single entry it will be fast.